homePage: all files renamed to .jsx + added prop-types
