+++ package.json
@@ -5,8 +5,10 @@
     "main": "index.js",
     "scripts": {
         "test": "echo \"Error: no test specified\" && exit 1",
-        "dev": "webpack --mode development ./src/index.js --output ./dist/main.js",
-        "build": "webpack --mode production ./src/index.js --output ./dist/main.js",
+        "dev":
+            "webpack --mode development ./src/index.jsx --output ./dist/main.js",
+        "build":
+            "webpack --mode production ./src/index.jsx --output ./dist/main.js",
         "watch": "webpack --watch",
         "start": "webpack-dev-server --open Chrome --mode development",
         "startOpt": "webpack-dev-server --open Chrome --mode production"
@@ -46,6 +48,7 @@
         "bootstrap": "^4.1.1",
         "font-awesome": "^4.7.0",
         "lodash": "^4.17.10",
+        "prop-types": "^15.6.2",
         "react": "^16.4.0",
         "react-dom": "^16.4.0",
         "react-router-dom": "^4.3.1"
