+++ src/app/applicationComponents/Footer/index.jsx
